Add back rack based server support
Add specs with rack server

Add workflow with rack present

Reduce test matrix size

Split spec files

Check rack presence purely based on gem

Remove with-rack matrix

Remove test starting rack server
