+++ Gemfile
@@ -10,5 +10,12 @@ gemspec
 gem "sidekiq", ENV["SIDEKIQ_VERSION_RANGE"] || "< 8"
 
 gem "ruby-lsp", "~> 0.5.1", group: :development
-gem "simplecov", require: false, group: :test
-gem "simplecov-cobertura"
+
+group :test do
+  gem "simplecov", require: false
+  gem "simplecov-cobertura"
+  # used for testing compatibility with rack based web server
+  gem "webrick", "< 2"
+  gem "rack", "< 3"
+  gem "rack-test", "~> 2.1.0"
+end
