+++ README.md
@@ -269,6 +269,15 @@ SidekiqAlive.setup do |config|
   # default: 2
   #
   #    config.concurrency = 3
+
+  # ==> Rack server
+  # Web server used to serve an HTTP response. By default simple GServer based http server is used.
+  # To use specific server, rack gem version < 3 is required.
+  # Can also be set with the environment variable SIDEKIQ_ALIVE_SERVER.
+  # default: nil
+  #
+  #    config.server = 'puma'
+
 end
 ```
